projecto/requirements.txt
Django
gunicorn
paho-mqtt
dj-database-url
django-database-url
psycopg2